+++ example-mobile/get_android_dependencies.ps1
@@ -1,4 +1,4 @@
-$AneVersion = "1.9.2"
+$AneVersion = "2.0.0"
 
 $currentDir = (Get-Item -Path ".\" -Verbose).FullName
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12